requirements.txt
Django==4.1.7
python-decouple==3.8
reportlab==4.0.9
daphne==2.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-sslserver==0.22
django-debug-toolbar==4.0.0
channels==2.4.0
psycopg2==2.9.5
Pillow==9.4.0
APScheduler==3.10.4
user-agents==2.2.0
requests==2.28.2
beautifulsoup4==4.12.2
redis==5.0.1
channels-redis==2.4.2
qrcode==7.4.2
stripe==7.5.0
xhtml2pdf==0.2.14
pypdf2==3.0.1